src/requirements.txt
aioredis==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.112.0
fastapi-cache2==0.2.2
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.0
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pendulum==3.0.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
redis==4.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
time-machine==2.15.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
websocket==0.2.1
websockets==13.0.1
zope.event==5.0
zope.interface==7.0.3
